scrapers updated

+++ crawlBitMidi.py
@@ -37,9 +37,15 @@ def scrape(data: dict, path: str = ''):
             fileName = os.path.join(path, key)
         else:
             fileName = key
-        print("Saving %s" % fileName)
+
+                # Check if the file already exists
+        fileName = os.path.join(path, key) if path else key
+        if os.path.exists(fileName + '.mid'):
+            print(f"File {fileName + '.mid'} already exists, skipping download.")
+            continue
 
         with open(fileName + '.mid', 'wb') as file:
+            print("Saving %s" % fileName)
             for chunk in midi.iter_content(chunk_size=1024):
                 # writing one chunk at a time
                 if chunk: 
@@ -87,8 +93,8 @@ if __name__ == "__main__":
             print(args.searchTerm)
         
         if args.path == None:
-            print("No path provided, saving to FreeMidi folder in current directory")
-            args.path = os.path.join(os.getcwd(), "FreeMidi")
+            print("No path provided, saving to BitMidi folder in current directory")
+            args.path = os.path.join(os.getcwd(), "BitMidi")
             if not os.path.exists(args.path):
                 os.makedirs(args.path)
         else:
@@ -116,7 +122,7 @@ if __name__ == "__main__":
                     f.write(jsonQuery)
             
         if args.download is True:
-            scrape(results)            
+            scrape(results, args.path)            
 else:
     print("running from import")
